fixing warnings

+++ yasli/decorators/BitFlags.cpp
@@ -10,7 +10,7 @@ void BitFlagsWrapper::serialize(yasli::Archive& ar)
 	if (ar.isInput()) {
 		//*variable = 0;
 		int previousValue = *variable;
-		for (size_t i = 0; i < count; ++i) {
+		for (size_t i = 0; i < (size_t)count; ++i) {
 			int flagValue = desc.valueByIndex(i);
 			bool flag = (previousValue & flagValue) == flagValue;
 			bool previousFlag = flag;
@@ -25,7 +25,7 @@ void BitFlagsWrapper::serialize(yasli::Archive& ar)
 		}
 	}
 	else {
-		for (size_t i = 0; i < count; ++i) {
+		for (size_t i = 0; i < (size_t)count; ++i) {
 			int flagValue = desc.valueByIndex(i);
 			bool flag = (*variable & flagValue) == flagValue;
 			ar(flag, desc.nameByIndex(i), desc.labelByIndex(i));
